+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 - Support for proxies
 - Support for sites that use basic-auth via the `.auth(username, password`)` API
+- New `refresh` API for reloading a page.
 - `.screenshot` takes a selector for the element to capture.
 
 ## Breaking